feat: add media queries
