+++ src/components/article/article.scss
@@ -8,23 +8,20 @@
     margin-bottom: 10px;
     word-wrap: break-word;
 
-    .content{
-        padding-left: 1em;
+    .iconLink{
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        padding-block-start: 1em;
+        color: #898989;
     }
 
-    .title{
-        color:$colorTitle;
+    @media(max-width: 640px){
+        .iconLink{
+            width: 15px;
+            height: 15px;
+            margin: 0 auto;
+        }
     }
 
-    .selfText{
-        color:  $colorContent;
-    }
-
-    .url{
-        color: $colorLink;
-        font-size: 13px;
-        font-weight: 600;
-        padding-bottom: 1em;
-        display: inline-block;
-    }
 }
