+++ src/components/article/article.tsx
@@ -1,27 +1,22 @@
 import React from 'react';
-import { ArticleInfo } from '../../shared/articleInfo';
+import { FaExternalLinkAlt } from 'react-icons/fa';
+import { InfoItem } from '../../shared/interfaces';
+import ContentArticle from '../contentArticle/contentArticle';
 import Likes from '../likes/likes';
 import './article.scss';
 
-interface InfoItem{
-  item: ArticleInfo
-}
-
 function Article({ item }:InfoItem):JSX.Element {
   const {
-    title, selftext, url, score,
+    score, url,
   } = item;
   return (
     <article className="article">
       <Likes score={score} />
-      <div className="content">
-        <h3 className="title">{title}</h3>
-        <p className="selfText">{selftext}</p>
-        <a className="url" href={url} target="_blank" rel="noreferrer">
-          {url.slice(0, 24)}
-          ...
-        </a>
-      </div>
+      <ContentArticle item={item} />
+      <a className="iconLink" href={url} target="_blank" rel="noreferrer">
+        <FaExternalLinkAlt size="20px" />
+      </a>
+
     </article>
   );
 }
