+++ src/components/contentArticle/contentArticle.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { InfoItem } from '../../shared/interfaces';
+import './contentArticle.scss';
+
+function ContentArticle({ item }:InfoItem):JSX.Element {
+  const {
+    title, selftext, url,
+  } = item;
+  return (
+    <div className="content">
+      <h3 className="title">{title}</h3>
+      <p className="selfText">{selftext}</p>
+      <a className="url" href={url} target="_blank" rel="noreferrer">
+        {url.slice(0, 24)}
+        ...
+      </a>
+    </div>
+  );
+}
+
+export default ContentArticle;
