+++ src/components/itemSidebar/itemSidebar.scss
@@ -7,6 +7,7 @@
     color: $colorSidebar;
     border: $border;
     background: white;
+    margin-bottom: 10px;
 
     .title{
         background: #0079d3;
@@ -27,5 +28,16 @@
         line-height: 21px;
         font-weight: 400;
         word-wrap: break-word;
+    }
+
+    @media(max-width: 460px){
+        .title{
+            font-size: 14px;
+        }
+
+        .content{
+            font-size: 12px;
+            line-height: 14px;
+        }
     }
 }
