--- src/shared/articleInfo.ts
@@ -1,7 +1,0 @@
-export interface ArticleInfo {
-    title: string
-    id: string
-    selftext: string
-    url: string
-    score: number
-}
